Move telebot into app/telebot using services directly
Replace HTTP calls to WEB_API, DB_API, and TELEMSG_API with direct
service/client calls:
- validate credentials: EvsClient().login()
- account/subscription CRUD: AccountService, SubscriptionService
- balance lookup: DbBalanceService
- command logging: CommandService
- admin/user messaging: context.bot.send_message()

Run with: python -m app.scripts.telebot
