+++ app/scripts/telebot.py
@@ -1,0 +1,6 @@
+from app.telebot.bot import Bot
+
+if __name__ == '__main__':
+    bot = Bot()
+    bot.add_handlers()
+    bot.start()
