+++ app/telebot/bot.py
@@ -1,13 +1,15 @@
 from telegram.ext import Updater
-from .handlers import commands as cmd, message
-from .handlers.conv import (add_subscription as add,
-                            view_subscription as view,
-                            feedback as feedback)
+
+from app.settings import TELEGRAM_TOKEN
+from app.telebot.handlers import commands as cmd, message
+from app.telebot.handlers.conv import (add_subscription as add,
+                                       view_subscription as view,
+                                       feedback)
 
 
-class Bot(object):
-    def __init__(self, token: str):
-        self.updater = Updater(token=token, use_context=True)
+class Bot:
+    def __init__(self):
+        self.updater = Updater(token=TELEGRAM_TOKEN, use_context=True)
         self.dispatcher = self.updater.dispatcher
 
     def add_handlers(self):
@@ -21,6 +23,3 @@ class Bot(object):
     def start(self):
         self.updater.start_polling()
         self.updater.idle()
-
-    def send_message(self, chat_id: int, text: str):
-        self.updater.bot.send_message(chat_id=chat_id, text=text)
