+++ app/telebot/handlers/conv/feedback.py
@@ -1,10 +1,9 @@
-import json
-import requests
-from telegram.ext import ConversationHandler, CommandHandler, MessageHandler, Filters
 from enum import Enum
 
-from ...config import TELEMSG_API_HOST, TELEMSG_API_PORT
-from ..logging import logger, log_command_in_db
+from telegram.ext import ConversationHandler, CommandHandler, MessageHandler, Filters
+
+from app.settings import TELEGRAM_ADMIN_ID
+from app.telebot.handlers.logging import logger, log_command_in_db
 
 COMMAND_NAME = 'feedback'
 
@@ -30,17 +29,14 @@ def feedback(update, context):
     chat_id = update.message.chat_id
     text = update.message.text
     logger.info(f'({chat_id}) Feedback: {text}')
+
     if text == '/cancel':
         return cancel(update, context)
 
-    message_successful = message_admin(chat_id, text)
-    if message_successful:
-        log_command_in_db(COMMAND_NAME, update.message.chat_id, is_completed=True, is_cancelled=False)
-        update.message.reply_text('Your message has been successfully received. Thank you!')
-    else:
-        log_command_in_db(COMMAND_NAME, update.message.chat_id, is_completed=False, is_cancelled=False)
-        update.message.reply_text('I am sorry, but your message could not be received.'
-                                  'Please try again later.')
+    context.bot.send_message(chat_id=int(TELEGRAM_ADMIN_ID),
+                             text=f'Message from {chat_id}:\n{text}')
+    log_command_in_db(COMMAND_NAME, chat_id, is_completed=True, is_cancelled=False)
+    update.message.reply_text('Your message has been successfully received. Thank you!')
     return ConversationHandler.END
 
 
@@ -48,14 +44,6 @@ def cancel(update, context):
     log_command_in_db(COMMAND_NAME, update.message.chat_id, is_completed=False, is_cancelled=True)
     update.message.reply_text('Ok bye')
     return ConversationHandler.END
-
-
-def message_admin(chat_id, text):
-    url = f'http://{TELEMSG_API_HOST}:{TELEMSG_API_PORT}/message/admin'
-    headers = {'Content-Type': 'application/json'}
-    data = json.dumps({'chat_id': chat_id, 'text': text})
-    req = requests.post(url, headers=headers, data=data)
-    return req.text == 'Success'
 
 
 conv_handler = ConversationHandler(
