--- hosts/telebot/.gitignore
@@ -1,2 +1,0 @@
-token.txt
-
