--- hosts/telebot/requirements.txt
@@ -1,2 +1,0 @@
-python-telegram-bot==13.5
-requests==2.25.1
